split upload and scoring into two files
